ACE-864: Import AMX BPM projects with Decisions Service assets in should remove them and any project config due to them
svn path=/trunk/; revision=89548

+++ tbs-bpm/product/com.tibco.xpd.n2.feature/plugins/com.tibco.xpd.n2.resources/src/com/tibco/xpd/n2/resources/postimport/Bpm2CeProjectConfigPostImportTask.java
@@ -360,14 +360,23 @@ public class Bpm2CeProjectConfigPostImportTask
 
                         /*
                          * Remove the physical folder itself if it exists.
+                         * Unless it's the decisions files as the user may not
+                         * realise what tehy're doing on import and might commit
+                         * back to svn without thinking (Jan requested in code
+                         * review)
+                         * 
                          */
-                        String sfLocation = specialFolder.getLocation();
+                        if (!"com.tibco.xpd.asset.decisions" //$NON-NLS-1$
+                                .equals(specialFolder.getKind())) {
+                            String sfLocation = specialFolder.getLocation();
 
-                        IFolder actualFolder = project.getFolder(sfLocation);
+                            IFolder actualFolder =
+                                    project.getFolder(sfLocation);
 
-                        if (actualFolder.exists()) {
-                            actualFolder.delete(true,
-                                    new NullProgressMonitor());
+                            if (actualFolder.exists()) {
+                                actualFolder.delete(true,
+                                        new NullProgressMonitor());
+                            }
                         }
 
                     }
